Reorder search for Java VM locations
We currently search for Apple's JVM first, and only if we don't find
that go look for the Oracle JVM. While that works, it has a major
downside: The Apple JVM is no longer supported by apple, and (on recent
versions of OS X) will issue a warning about it being an older version
of Java.

If we have code that specifies it will work with any version of Java, it
is usually best to use the most recent JVM rather than the oldest one.
Since Oracle's JVMs are actually maintained these days, defaulting to
that seems like a better strategy than defaulting to Apple's JVM.