Allow specifying `JAVA_HOME` relative to `$AppPackageFolder`
One may wish, for example, to bundle the JRE with their application.  When doing so, it is helpful to specify the `JAVA_HOME` relative to the application package folder.  This patch will prepend `$AppPackageFolder` to `JAVA_HOME` when looking for `JAVACMD` - but *ONLY* if `JAVA_HOME` is a relative path (doesn't start with `/`).  This allows you to set a relative `JAVA_HOME` using `<lsenvironment>` and it will get picked up.