Revert "Pass command line arguments through to the application only if they exist"
This reverts commit 4105979702bc685508326c5d168ef70728ac2f18.

It turns out "open" sometimes passes a "process serial number" argument of the form -psn_X_YYYYYYY to apps when launching them. What little documentation I could find suggests the behavior differs across OS versions and whether or not the app is already open, so it's probably best not to worry about it in universalJavaApplicationStub and let the application handle it if present. Also, use "$@" instead of $@ to preserve argument whitespace.

For my app, on MacOS 10.10.5, the extra argument appears to be passed only if open is used without the --args option. My app terminates with a help message when such an unexpected argument is received, but open redirects the output to somewhere I can't find, so the observed behavior was silent termination after the splash screen.